refactor(main): expose IO modules
