+++ package.json
@@ -23,7 +23,7 @@
   "engines": {
     "node": ">= 5.0.0"
   },
-  "main": ".dist/index.js",
+  "main": ".dist/Main.js",
   "dependencies": {
     "graceful-fs": "^4.1.3",
     "humanize-plus": "^1.8.1",
